Renamed building files to hotel
