+++ lib/hotel.rb
