+++ specs/hotel_spec.rb
