+++ specs/spec_helper.rb
@@ -8,7 +8,7 @@ require 'minitest/skip_dsl'
 require 'minitest/pride'
 require_relative '../lib/room'
 require_relative '../lib/reservation'
-require_relative '../lib/building'
+require_relative '../lib/hotel'
 
 reporter_options = { color:true}
 Minitest::Reporters.use! Minitest::Reporters::DefaultReporter.new(reporter_options)
